Added module:HasTimer(name) to the core (NEED QA BEFORE IMPLEMENTING)
Modules should be inheriting this function.
This Function should check if a Timer/Bar with 'name' is already being
displayed to prevent in some cases multiple bars being shown at the same
time in case the SYNC doesn't work correctly.

THIS NEEDS QA BADLY BEFORE IMPLEMENTING.
This was just a quick edit, I'm not at home and can't test it right
away.