removing Throttle for BossEngage Sync
the reason being - we are checking in each bossmod for self.started and
prevent it like that to call the initial timers multiple times.

I left the old engage sync in it so that ppl with older bigwigs version
in the group will also receive the sync they can handle.